Write set(key, value) function
This function takes a key and a value, generates a hash value(index for the bucket in
the hashtable), creates a linkedList in the generated index if there's none yet and
appends the key/value pair passed. If there's already a linkedList(collision), it appends the
key/value pair to the end of the list. Finally if the key passed is already present in
any bucket, it only updates the value of it, since the keys are the same.

After every insertion, we check if the number of buckets in the hashtable is above
the loadfactor, if it is we duplicate the size of the hashtable and rehash the keys to
insert them in new bucket indexes.